Buy tab is extracted as a separate component
